Delete Demandes AJAX

+++ Plannr/Controllers/BookController.cs
@@ -15,11 +15,13 @@ using System.Web.Script.Serialization;
 
 namespace Plannr.Controllers
 {
+      [InitializeSimpleMembership]
     public class BookController : Controller
     {   
         // Repository as private member, generic interface type
         private IDemandesRepository repository;
         private IEnseignementsRepository enseignementsRepository;
+        private ICreneauxHorairesRepository creneauxHorairesRepository;
 
         // Constructor
         public BookController()
@@ -28,13 +30,16 @@ namespace Plannr.Controllers
             var context = new PlannrContext();
             this.repository = new DemandesRepository(context);
             this.enseignementsRepository = new EnseignementsRepository(context);
+            this.creneauxHorairesRepository = new CreneauxHorairesRepository(context);
+
         }
 
         // Give it as a parameter aswel
-        public BookController(IDemandesRepository repo, IEnseignementsRepository ensRepo)
+        public BookController(IDemandesRepository repo, IEnseignementsRepository ensRepo, ICreneauxHorairesRepository crRepo)
         {
             this.repository = repo;
             this.enseignementsRepository = ensRepo;
+            this.creneauxHorairesRepository = crRepo;
         }
 
 
@@ -47,18 +52,21 @@ namespace Plannr.Controllers
             var id = (int) Membership.GetUser().ProviderUserKey;
             IEnumerable<DemandeReservation> demandes = this.repository.GetReservationsBy(id).ToList();
 
-            var data = demandes.Select(x => new
+           /* var data = demandes.Select(x => new
             {
                 DateDemande = x.DateDemande.ToString(),
                 DateVoulue = x.DateVoulue.ToString(),
-                Enseignement = x.Enseignement.Cours.Libelle+" - "+x.Enseignement.Groupe.Libelle,
-                Creneau = x.HeureDepart+" - "+x.HeureFin,
-                Capacite = x.CapaciteNecesaire
-
+                Enseignement = x.Enseignement_Libelle,
+                Creneau = x.CreneauSouhaite.HeureConcat,
+                Capacite = x.CapaciteNecessaire,
+                Reservation_Id = x.ReservationAssociee_Id
            
-            });
+            });*/
 
-            ViewBag.demandes = new JavaScriptSerializer().Serialize(Json(data).Data);
+            JsonSerializerSettings jsSettings = new JsonSerializerSettings();
+            jsSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
+
+            ViewBag.JSONModel = JsonConvert.SerializeObject(demandes, Formatting.None, jsSettings);
 
             // Render full or partial.
             if (!Request.IsAjaxRequest())
@@ -83,6 +91,7 @@ namespace Plannr.Controllers
             var id = (int)Membership.GetUser().ProviderUserKey;
             // On va chercher les enseignements qui sont attribués à l'enseignant pour les lister dans la DropList en sélectionner un, on le passe à la vue pour génrer la ListBox
             ViewBag.listEnseignements = this.enseignementsRepository.GetEnseignementsForTeacher(id);
+            ViewBag.listCreneauxHoraires = this.creneauxHorairesRepository.getCreneauxHoraires();
                                     
             return View();
         }
@@ -95,40 +104,46 @@ namespace Plannr.Controllers
             
             demandereservation.DateDemande = DateTime.Now;
             demandereservation.Checked = false;
+            
             // Mapping
             demandereservation.Enseignement = this.enseignementsRepository.Get(demandereservation.Enseignement.Id);
+            demandereservation.CreneauSouhaite = this.creneauxHorairesRepository.Find(demandereservation.CreneauSouhaite.Id);
+      
 
             if (ModelState.IsValid)
             {
                 repository.Insert(demandereservation);
                 repository.Save();
-              
+
                 return RedirectToAction("Index");
             }
+
+            ModelState.Values.Where(modelState => modelState.Errors.Count > 0).ToList().ForEach(x => x.Errors.ToList().ForEach(y => System.Diagnostics.Debug.WriteLine(y.ErrorMessage)));
 
 
             // Since it's returing view, we need the listEnseignements in the ViewBag aswell
             var id = (int)Membership.GetUser().ProviderUserKey;
             ViewBag.listEnseignements = this.enseignementsRepository.GetEnseignementsForTeacher(id);
+            ViewBag.listCreneauxHoraires = this.creneauxHorairesRepository.getCreneauxHoraires();
 
  
 
             return View(demandereservation);
         }
 
-  /*
+  
         //
         // GET: /Book/Delete/5
         [Authorize(Roles = "Enseignant")]
         public ActionResult Delete(int id = 0)
         {
-            DemandeReservation demandereservation = db.DemandesReservation.Find(id);
+            DemandeReservation demandereservation = this.repository.Find(id);
             if (demandereservation == null)
             {
                 return HttpNotFound();
             }
             return View(demandereservation);
-        }*/
+        }
 
         //
         // POST: /Book/Delete/5
@@ -136,9 +151,36 @@ namespace Plannr.Controllers
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
+            var demande = this.repository.Find(id);
+
+            if (demande.ReservationAssociee != null)
+            {
+                // On ne supprime pas une demande qui a déjà été validée.
+                return new HttpStatusCodeResult(403);
+            }
+
+            // On vérifie que la demande existe.
+            if (demande == null)
+            {
+                return new HttpStatusCodeResult(404);
+            }
+
+            var SessionId = (int)Membership.GetUser().ProviderUserKey;
+            if (demande.Enseignement.Enseignant.UserId != SessionId)
+            {
+                return new HttpStatusCodeResult(403);
+            }
+
             this.repository.Delete(id);
             this.repository.Save();
-            return RedirectToAction("Index");
+            if (!Request.IsAjaxRequest())
+            {
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return new HttpStatusCodeResult(200);
+            }
         }
 
     }
